+++ .php_cs
@@ -15,7 +15,7 @@ $config = Config\Factory::fromRuleSet(new Config\RuleSet\Php56($header));
 
 $config->getFinder()
     ->in(__DIR__)
-    ->exclude('test/Unit/Asset/Definition/CanNotBeAutoloaded');
+    ->exclude('test/Fixture/Definition/CanNotBeAutoloaded');
 
 $cacheDir = \getenv('TRAVIS') ? \getenv('HOME') . '/.php-cs-fixer' : __DIR__;
