+++ test/Fixture/Definition/ThrowsExceptionDuringConstruction/UserDefinition.php
@@ -9,22 +9,23 @@
  * @link https://github.com/localheinz/factory-girl-definition
  */
 
-namespace Localheinz\FactoryGirl\Definition\Test\Unit\Asset\Definition\PrivateConstructor;
+namespace Localheinz\FactoryGirl\Definition\Test\Fixture\Definition\ThrowsExceptionDuringConstruction;
 
 use FactoryGirl\Provider\Doctrine\FixtureFactory;
 use Localheinz\FactoryGirl\Definition\Definition;
 
 /**
- * Is not acceptable as it has a private constructor.
+ * Is not acceptable as it throws an exception during construction.
  */
 final class UserDefinition implements Definition
 {
-    private function __construct()
+    public function __construct()
     {
+        throw new \RuntimeException();
     }
 
     public function accept(FixtureFactory $factory)
     {
-        $factory->defineEntity('Foo');
+        $factory->defineEntity(\Localheinz\FactoryGirl\Definition\Test\Fixture\Entity\User::class);
     }
 }
